Fixed: load timed quest timer as time_t value. This fix possible unexpected timed quest fail at character load. Fixed: restore support build mangos at Unix/Linux in case build_dir != src_dir Fixed: lot warnings output at compile using GCC.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@4130 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Group.cpp
@@ -356,7 +356,7 @@ void Group::GroupLoot(uint64 playerGUID, Loot *loot, Creature *creature)
             //sLog.outDebug("Group::GroupLoot: missing item prototype for item with id: %d", i->itemid);
             continue;
         }
-        if (item->Quality >= m_lootThreshold)
+        if (item->Quality >= uint32(m_lootThreshold))
         {
             Roll r;
             uint32 newitemGUID = objmgr.GenerateLowGuid(HIGHGUID_ITEM);
@@ -404,7 +404,7 @@ void Group::NeedBeforeGreed(uint64 playerGUID, Loot *loot, Creature *creature)
     for (i=loot->items.begin(); i != loot->items.end(); i++)
     {
         item = objmgr.GetItemPrototype(i->itemid);
-        if (item->Quality >= m_lootThreshold)
+        if (item->Quality >= uint32(m_lootThreshold))
         {
             Roll r;
             uint32 newitemGUID = objmgr.GenerateLowGuid(HIGHGUID_ITEM);
